Issue in IndexManager.getIndexer()
The implementation class of the indexer is created only once in the
static initializer block, instead of every time when getIndexer() is
called. This makes the code more robust of eventual issues when loading
the implementation class.

If there is a CoreException thrown, it is always logged in the error
log.

Bug: 422573
Signed-off-by: Kaloyan Raev <kaloyan.r@zend.com>